requirements.txt
chromadb==0.4.24
rank_bm25==0.2.2
flash-attn==2.5.8
protobuf==4.25.2
sentencepiece==0.2.0
peft==0.11.1
beautifulsoup4==4.12.3
langchain==0.2.1
langchain_community==0.2.1
langchain_core==0.2.3
langchain_huggingface==0.0.1
langchain_openai==0.1.8
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
tqdm==4.66.1
transformers==4.41.1
fastapi==0.111.0
snowballstemmer==2.2.0
razdel==0.5.0
pydantic
stop-words